move constants to constants.ts
