--- project/src/components/app/routes.ts
@@ -1,8 +1,0 @@
-export const enum AppRoutes {
-  Main = '/',
-  SignIn = '/login',
-  MyList = '/mylist',
-  Films = '/films/:id',
-  Review = '/films/:id/review',
-  Player = '/player/:id',
-}
